requirements.txt
requests==2.31.1
PyGithub==1.58.3
keyboard==0.13.5